+++ crates/shared/src/model/topup.rs
@@ -7,7 +7,7 @@ pub struct TopupModel {
     pub id: i32,
     pub card_number: String,
     pub topup_no: String,
-    pub topup_amount: i32,
+    pub topup_amount: i64,
     pub topup_method: String,
     pub topup_time: String,
     pub created_at: Option<NaiveDateTime>,
@@ -20,14 +20,14 @@ pub struct TopupModelMonthStatusSuccess {
     pub year: String,
     pub month: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TopupModelYearStatusSuccess {
     pub year: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -35,14 +35,14 @@ pub struct TopupModelMonthStatusFailed {
     pub year: String,
     pub month: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TopupModelYearStatusFailed {
     pub year: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -50,7 +50,7 @@ pub struct TopupMonthMethod {
     pub month: String,
     pub topup_method: String,
     pub total_topups: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -58,17 +58,17 @@ pub struct TopupYearlyMethod {
     pub year: String,
     pub topup_method: String,
     pub total_topups: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TopupMonthAmount {
     pub month: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TopupYearlyAmount {
     pub year: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
