+++ crates/shared/src/model/transaction.rs
@@ -7,7 +7,7 @@ pub struct TransactionModel {
     pub id: i32,
     pub card_number: String,
     pub transaction_no: String,
-    pub amount: i32,
+    pub amount: i64,
     pub payment_method: String,
     pub merchant_id: i32,
     pub transaction_time: String,
@@ -21,14 +21,14 @@ pub struct TransactionModelMonthStatusSuccess {
     pub year: String,
     pub month: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransactionModelYearStatusSuccess {
     pub year: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -36,14 +36,14 @@ pub struct TransactionModelMonthStatusFailed {
     pub year: String,
     pub month: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransactionModelYearStatusFailed {
     pub year: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -51,7 +51,7 @@ pub struct TransactionMonthMethod {
     pub month: String,
     pub payment_method: String,
     pub total_transactions: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -59,17 +59,17 @@ pub struct TransactionYearMethod {
     pub year: String,
     pub payment_method: String,
     pub total_transactions: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransactionMonthAmount {
     pub month: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransactionYearlyAmount {
     pub year: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
