+++ crates/shared/src/model/transfer.rs
@@ -8,7 +8,7 @@ pub struct TransferModel {
     pub transfer_no: String,
     pub transfer_from: String,
     pub transfer_to: String,
-    pub transfer_amount: i32,
+    pub transfer_amount: i64,
     pub transfer_time: String,
     pub created_at: Option<NaiveDateTime>,
     pub updated_at: Option<NaiveDateTime>,
@@ -20,14 +20,14 @@ pub struct TransferModelMonthStatusSuccess {
     pub year: String,
     pub month: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransferModelYearStatusSuccess {
     pub year: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -35,24 +35,24 @@ pub struct TransferModelMonthStatusFailed {
     pub year: String,
     pub month: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransferModelYearStatusFailed {
     pub year: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransferMonthAmount {
     pub month: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct TransferYearAmount {
     pub year: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
