+++ crates/shared/src/model/withdraw.rs
@@ -7,7 +7,7 @@ pub struct WithdrawModel {
     pub id: i32,
     pub withdraw_no: String,
     pub card_number: String,
-    pub withdraw_amount: i32,
+    pub withdraw_amount: i64,
     pub withdraw_time: NaiveDateTime,
     pub created_at: Option<NaiveDateTime>,
     pub updated_at: Option<NaiveDateTime>,
@@ -19,14 +19,14 @@ pub struct WithdrawModelMonthStatusSuccess {
     pub year: String,
     pub month: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct WithdrawModelYearStatusSuccess {
     pub year: String,
     pub total_success: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
@@ -34,24 +34,24 @@ pub struct WithdrawModelMonthStatusFailed {
     pub year: String,
     pub month: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct WithdrawModelYearStatusFailed {
     pub year: String,
     pub total_failed: i32,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct WithdrawMonthlyAmount {
     pub month: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize, FromRow)]
 pub struct WithdrawYearlyAmount {
     pub year: String,
-    pub total_amount: i32,
+    pub total_amount: i64,
 }
